added --full option

+++ shutdown.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 docker-compose -p sql_fun down
-rm -rf pgdata
-docker volume prune --force 
-docker network prune --force 
+
+# Remove all changes
+if [ "$1" == "--full" ]; then
+  docker volume prune --force 
+  docker network prune --force 
+  # removing postgres data
+  rm -rf pgdata
+fi
